requirements.txt
# %%writefile /content/drive/MyDrive/DeforestationApp/requirements.txt
torch
torchvision
timm
rasterio
numpy
pillow
gradio